// backend/src/types/resume.ts
export interface Resume {
    id: string;
    name: string;
    email: string;
    phone: string;
    skills: string[];
    experience: Experience[];
    education: Education[];
  }
  
  export interface Experience {
    company: string;
    position: string;
    startDate: string;
    endDate: string;
    description: string[];
  }
  
  export interface Education {
    institution: string;
    degree: string;
    field: string;
    graduationDate: string;
  }